+++ src/components/FormSection.vue
@@ -78,9 +78,29 @@ export default {
       'isZipFieldExposed',
       'isZipInvalid',
       'currentPlatformId',
+      // Todo: remove after #195691
+      'hasExpAutofill',
     ]),
-    ...mapGetters('PaymentForm', ['activePaymentMethod']),
+    ...mapGetters('PaymentForm', [
+      'activePaymentMethod',
+      'isPaymentFailed',
+      'isPaymentSuccess',
+    ]),
+    ...mapGetters('PaymentForm/Redirect', [
+      'isRedirect',
+      'isAutoRedirect',
+      'redirectDelay',
+      'hasTimer',
+      'redirectSuccessUrl',
+      'redirectFailUrl',
+    ]),
     ...mapGetters('Dictionaries', ['countries']),
+
+    redirectButtonText() {
+      return this.isRedirect
+        ? get(this.orderData, 'project.redirect_settings.button_caption', '')
+        : '';
+    },
 
     paymentMethodsSelectList() {
       return this.orderData.payment_methods.map((item) => {
@@ -130,14 +150,6 @@ export default {
 
     platformInstructionLink() {
       return geInstructionLinkByPlatform(this.currentPlatformId);
-    },
-
-    isPaymentSuccess() {
-      return this.actionResult && this.actionResult.type === 'success';
-    },
-
-    isPaymentFailed() {
-      return this.actionResult && this.actionResult.type !== 'success';
     },
   },
 
@@ -196,19 +208,27 @@ export default {
         if (this.$refs.bankCardForm) {
           this.$refs.bankCardForm.focusCardNumberField();
         }
-      } else if (this.isUserCountryRestricted || this.isPaymentFailed) {
+      } else if (this.isUserCountryRestricted) {
         gtagEvent('clickCloseButton', { event_category: 'userAction' });
-        this.$emit('close');
-      } else if (this.isPaymentSuccess) {
-        gtagEvent('clickOkButton', { event_category: 'userAction' });
         this.$emit('close');
       } else if (this.isPaymentFormVisible) {
         gtagEvent('clickPayButton', { event_category: 'userAction' });
         this.submitPaymentForm();
+      } else if (this.isRedirect) {
+        const eventName = this.isPaymentFailed ? 'clickCloseButton' : 'clickOkButton';
+        const redirectUrl = this.isPaymentFailed ? this.redirectFailUrl : this.redirectSuccessUrl;
+
+        gtagEvent(eventName, { event_category: 'userAction' });
+
+        if (this.isAutoRedirect === false) {
+          window.location.replace(redirectUrl);
+        }
+      } else {
+        this.$emit('close');
       }
     },
 
-    async submitPaymentForm() {
+    submitPaymentForm() {
       this.$v.$touch();
 
       const isValidArray = [
@@ -226,6 +246,7 @@ export default {
       }
 
       gtagEvent('submitPaymentForm');
+
       this.createPayment({
         ...this.paymentData,
       });
@@ -303,6 +324,7 @@ export default {
           :isCountryFieldExposed="isCountryFieldExposed"
           :isZipFieldExposed="isZipFieldExposed"
           :isZipInvalid="isZipInvalid"
+          :hasExpAutofill="hasExpAutofill"
           @savedCardIdChange="checkSavedCardNumberById"
           @cardNumberChange="checkBankCardNumber"
           @countryChange="setNewUserCountry($event), requestBillingDataUpdate()"
@@ -357,7 +379,16 @@ export default {
     </component>
   </div>
   <div :class="$style.footer">
+    <UiTimer
+      v-if="hasTimer"
+      minWidth="100px"
+      :time="redirectDelay"
+    >
+      <IconTimer slot="prepend" />
+      <span slot="append">{{ $t('FormSection.seconds') }}</span>
+    </UiTimer>
     <UiButton
+      v-else
       :class="$style.payBtn"
       :hasBorderRadius="isPageView"
       :disabled="isSubmitButtonDisabled"
@@ -372,14 +403,17 @@ export default {
           {{ $getPrice(orderData.charge_amount, orderData.charge_currency) }}
         </span>
       </template>
-      <template v-if="isPaymentSuccess">
-        {{ $t('FormSection.ok') }}
-      </template>
       <template v-if="isUserCountryConfirmRequested">
         {{ $t('FormSection.save') }}
       </template>
-      <template v-if="isUserCountryRestricted || isPaymentFailed">
+      <template v-if="isUserCountryRestricted">
         {{ $t('FormSection.close') }}
+      </template>
+      <template v-if="isPaymentFailed">
+        {{ redirectButtonText || $t('FormSection.close') }}
+      </template>
+      <template v-if="isPaymentSuccess">
+        {{ redirectButtonText || $t('FormSection.ok') }}
       </template>
     </UiButton>
   </div>
@@ -397,7 +431,6 @@ export default {
   max-height: 100%;
   width: 100%;
 }
-
 .content {
   display: flex;
   flex-wrap: wrap;
@@ -408,13 +441,11 @@ export default {
   width: 100%;
   max-height: calc(100% - 70px);
 }
-
 .scrollbox {
   width: 100%;
   height: 100%;
   flex-grow: 1;
 }
-
 .contentInner {
   width: 100%;
   height: 100%;
@@ -429,15 +460,13 @@ export default {
     }
   }
 }
-
 .formItem {
   display: flex;
   justify-content: space-between;
 }
-
 .footer {
   display: flex;
-  justify-content: flex-start;
+  justify-content: center;
   align-items: flex-end;
   width: 100%;
 
@@ -446,7 +475,6 @@ export default {
     width: 100vw !important;
   }
 }
-
 .payBtn {
   width: 100%;
   transition: border-radius 0.2s ease-out;