+++ src/viewSchemes.js
@@ -124,6 +124,13 @@ export default {
     resultEmailColor: '#ffffff',
 
     scrollbarColor: 'rgba(0, 0, 0, 0.7)',
+
+    timerColor: '#f255ed',
+    timerBorderColor: '#ffffff',
+    timerShadowColor: '#333b50',
+
+    testNotifyBackgroundColor: 'rgba(243, 170, 24, 0.2)',
+    testNotifyColor: '#f3aa18',
   },
   light: {
     baseHoverColor: '#06eaa7',
@@ -250,5 +257,12 @@ export default {
     resultEmailColor: '#202226',
 
     scrollbarColor: 'rgba(0, 0, 0, 0.7)',
+
+    timerColor: '#f255ed',
+    timerBorderColor: '#ffffff',
+    timerShadowColor: '#333b50',
+
+    testNotifyBackgroundColor: 'rgba(243, 170, 24, 0.2)',
+    testNotifyColor: '#f3aa18',
   },
 };